refactor(analysis): complete module migration
